// src/common/Enum.ts
export enum RCode {
    success = 200,
    created = 201,
    accepted = 202,
    found = 302,
    badRequest = 400,
    unAuthorized = 401,
    forbidden = 403,
    notFound = 404,
    notAcceptable = 406,
    conflict = 409,
    expectationFailed = 417,
    internalServerError = 500,
    serviceUnAvailable = 503
}